Fix olympus and uspce report tests
